py/requirements.txt
pyserial
pywin32